+++ changes.md
@@ -2,6 +2,7 @@
 
 ## 21 June 2023
 - Automatically create new Additional Style Files if needed
+- Automatically delete empty Additional Style Files on merge
 
 ## 20 June 2023
 - Regular backups created in `extensions/Styles-Editor/backups` (saved every ten minutes if changes have been made, last 12 retained)